reorganise state schema and prepare for refactor
