+++ kvm-compose/kvm-compose/src/exec/file_transfer.rs
@@ -7,7 +7,7 @@ use crate::components::helpers::serialisation;
 use crate::exec::libvirt::shell_command;
 use crate::orchestration::api::OrchestrationLogger;
 use crate::orchestration::OrchestrationCommon;
-use crate::state::StateTestbedGuest;
+use crate::state::schema::StateTestbedGuest;
 
 const CDROM_DEVICE_XML: &str = r#"
     <disk type='file' device='cdrom'>
