+++ kvm-compose/kvm-compose/src/state/orchestration_tasks/mod.rs
@@ -3,7 +3,7 @@ use anyhow::{bail, Context};
 use async_trait::async_trait;
 use futures_util::future::try_join_all;
 use nix::unistd::{Gid, Uid};
-use tokio::sync::mpsc::{Sender};
+use tokio::sync::mpsc::Sender;
 use kvm_compose_schemas::kvm_compose_yaml::machines::GuestType;
 use kvm_compose_schemas::kvm_compose_yaml::machines::libvirt::LibvirtGuestOptions;
 use kvm_compose_schemas::settings::TestbedClusterConfig;
@@ -13,9 +13,9 @@ use crate::components::LogicalTestbed;
 use crate::get_project_folder_user_group;
 use crate::orchestration::*;
 use crate::orchestration::api::{OrchestrationInstruction, OrchestrationLogger, OrchestrationProtocol};
-use crate::orchestration::websocket::{send_orchestration_instruction_over_channel};
-use crate::state::{State, StateTestbedGuestList};
+use crate::orchestration::websocket::send_orchestration_instruction_over_channel;
 use crate::state::orchestration_tasks::stages::*;
+use crate::state::schema::{State, StateTestbedGuestList};
 
 pub mod ovs_network;
 pub mod ovn_network;
